requirements.txt
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flasgger==0.9.7.1
